HPCC-34115: Decouple event consumption from jlib, part 1
- Create a new shared library for event consumption logic.
  - Include all event dump visitors.
  - Include the event filter.
  - Include base operation class implementing `traverseEvents`. This allows
    `readEvents`, which remains in jlib, to not be aware of event filters.
- Update jevent in jlib.
  - Remove all event dump visitors.
  - Remove event filters.
  - Expose event attribute data types.
  - Expose mapping of event context name to context.
  - Both `readEvents` and `IEventVisitor` are retained, for now.
- Update evtool and unittests to use new library.
