+++ common/eventconsumption/eventvisitor.h
@@ -1,0 +1,23 @@
+/*##############################################################################
+
+    Copyright (C) 2025 HPCC Systems®.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+############################################################################## */
+
+#pragma once
+
+// MORE: If jevent.hpp replaces readEvents with a function that returns a reader instance, from
+// which events are pulled instead of visisted, then this file will declare the IEventVisisot
+// interface and the adapter to convert pulled events into visisted events.
+#include "jevent.hpp"
