+++ system/jlib/jevent.cpp
@@ -100,21 +100,6 @@ static constexpr EventInformation eventInformation[] {
 };
 static_assert(_elements_in(eventInformation) == EventMax);
 
-// The different types of attributes that can be associated with an event.
-enum EventAttrType
-{
-    EATnone,
-    EATbool,
-    EATu1,
-    EATu2,
-    EATu4,
-    EATu8,
-    EATtimestamp,
-    EATstring,
-    EATtraceid,     // should probably delete...
-    EATmax
-};
-
 static constexpr unsigned attrTypeSizes[] = { 0, 1, 1, 2, 4, 8, 8, 0, 32 };
 static_assert(_elements_in(attrTypeSizes) == EATmax);
 
@@ -198,6 +183,12 @@ const char * queryEventAttributeName(EventAttr attr)
 {
     assertex(attr < EvAttrMax);
     return attrInformation[attr].name;
+}
+
+EventAttrType queryEventAttributeType(EventAttr attr)
+{
+    assertex(attr < EvAttrMax);
+    return attrInformation[attr].type;
 }
 
 //---------------------------------------------------------------------------------------------------------------------
@@ -1108,1191 +1099,10 @@ private:
     }
 };
 
-bool readEvents(const char* filename, IEventVisitor& visitor, IEventFilter* filter)
+bool readEvents(const char* filename, IEventVisitor& visitor)
 {
     CEventFileReader reader;
-    if (filter)
-    {
-        filter->setTarget(visitor);
-        return reader.traverse(filename, *filter);
-    }
     return reader.traverse(filename, visitor);
-}
-
-// Event data may be expressed in multiple structured formats, such as XML, JSON, or YAML. Each of
-// these can be represented as in an IPropertyTree. All such representations should share a common
-// structure.
-//
-// The IPropertyTree representation of a binary event file resembles:
-//     EventFile
-//       ├── Header
-//       │   ├── @filename
-//       │   ├── @version
-//       │   └── ... (other header attributes)
-//       ├── Event
-//       │   ├── @name
-//       │   └── ... (other event attributes)
-//       ├── ... (more Event elements)
-//       └── Footer
-//           ├── @bytesRead
-
-constexpr static const char* DUMP_STRUCTURE_ROOT = "EventFile";
-constexpr static const char* DUMP_STRUCTURE_FILE_NAME = "filename";
-constexpr static const char* DUMP_STRUCTURE_HEADER = "Header";
-constexpr static const char* DUMP_STRUCTURE_FILE_VERSION = "version";
-constexpr static const char* DUMP_STRUCTURE_EVENT = "Event";
-constexpr static const char* DUMP_STRUCTURE_EVENT_NAME = "name";
-constexpr static const char* DUMP_STRUCTURE_FOOTER = "Footer";
-constexpr static const char* DUMP_STRUCTURE_FILE_BYTES_READ = "bytesRead";
-
-// Abstract base class for event visitors that store visited data in a text-based format. Text-
-// based, in this context, means only that the values will be stored as strings. The actual output
-// format is determined by the subclass.
-//
-// Subclasses must implement:
-//  `virtual bool visitFile(const char* filename, uint32_t version) = 0;`
-//  `virtual Continuation visitEvent(EventType id) = 0;`
-//  `virtual bool departEvent() = 0;`
-//  `virtual void departFile(uint32_t bytesRead) = 0;`
-//  `virtual void recordAttribute(EventAttr id, const char* name, const char* value, bool quoted) = 0;`
-class CDumpEventVisitor : public CInterfaceOf<IEventVisitor>
-{
-public:
-    using Continuation = IEventVisitor::Continuation;
-
-    virtual Continuation visitAttribute(EventAttr id, const char* value) override
-    {
-        if (EvAttrRecordedOption == id)
-            recordAttribute(EvAttrNone, value, "true", false);
-        else
-            doVisitAttribute(id, value);
-        return Continuation::visitContinue;
-    }
-
-    virtual Continuation visitAttribute(EventAttr id, bool value) override
-    {
-        doVisitAttribute(id, value);
-        return Continuation::visitContinue;
-    }
-
-    virtual Continuation visitAttribute(EventAttr id, __uint64 value) override
-    {
-        doVisitAttribute(id, value);
-        return Continuation::visitContinue;
-    }
-
-protected:
-    void doVisitHeader(const char* filename, uint32_t version)
-    {
-        doVisitAttribute(EvAttrNone, DUMP_STRUCTURE_FILE_NAME, filename);
-        doVisitAttribute(EvAttrNone, DUMP_STRUCTURE_FILE_VERSION, __uint64(version));
-    }
-
-    void doVisitEvent(EventType id)
-    {
-        doVisitAttribute(EvAttrNone, DUMP_STRUCTURE_EVENT_NAME, queryEventName(id));
-    }
-
-    void doVisitAttribute(EventAttr id, __uint64 value)
-    {
-        doVisitAttribute(id, queryEventAttributeName(id), value);
-    }
-
-    void doVisitAttribute(EventAttr id, const char* value)
-    {
-        doVisitAttribute(id, queryEventAttributeName(id), value);
-    }
-
-    void doVisitAttribute(EventAttr id, bool value)
-    {
-        doVisitAttribute(id, queryEventAttributeName(id), value);
-    }
-
-    void doVisitAttribute(EventAttr id, const char* name, __uint64 value)
-    {
-        if (attrInformation[id].type == EATtimestamp)
-        {
-            StringBuffer timestamp;
-            CDateTime dt;
-            dt.setTimeStampNs(value);
-            dt.getString(timestamp);
-            // assumes CDateTime output is in microseconds
-            timestamp.appendf("%03llu", value % 1000);
-            recordAttribute(id, name, timestamp, true);
-            return;
-        }
-        recordAttribute(id, name, StringBuffer().append(value), false);
-    }
-
-    void doVisitAttribute(EventAttr id, const char* name, const char* value)
-    {
-        recordAttribute(id, name, value, true);
-    }
-
-    void doVisitAttribute(EventAttr id, const char* name, bool value)
-    {
-        recordAttribute(id, name, (value ? "true" : "false"), false);
-    }
-
-    void doVisitFooter(uint32_t bytesRead)
-    {
-        doVisitAttribute(EvAttrNone, DUMP_STRUCTURE_FILE_BYTES_READ, __uint64(bytesRead));
-    }
-
-    virtual void recordAttribute(EventAttr id, const char* name, const char* value, bool quoted) = 0;
-};
-
-// Concrete extension of CDumpEventVisitor<IPTreeEventVisitor> that uses an IPropertyTree to store
-// visited data. The tree is constructed in memory and is available via the `queryTree` method
-// after visitation.
-class CPTreeEventVisitor : public CDumpEventVisitor
-{
-public: // IEventVisitor
-    virtual bool visitFile(const char* filename, uint32_t version) override
-    {
-        tree.setown(createPTree(DUMP_STRUCTURE_ROOT));
-        active = tree->addPropTree(DUMP_STRUCTURE_HEADER, createPTree());
-        doVisitHeader(filename, version);
-        return true;
-    }
-
-    virtual Continuation visitEvent(EventType id) override
-    {
-        active = tree->addPropTree(DUMP_STRUCTURE_EVENT, createPTree());
-        doVisitEvent(id);
-        return visitContinue;
-    }
-
-    virtual bool departEvent() override
-    {
-        active = tree.get();
-        return true;
-    }
-
-    virtual void departFile(uint32_t bytesRead) override
-    {
-        active = tree->addPropTree(DUMP_STRUCTURE_FOOTER, createPTree());
-        doVisitFooter(bytesRead);
-    }
-
-protected:
-    virtual void recordAttribute(EventAttr id, const char* name, const char* value, bool quoted) override
-    {
-        active->addProp(VStringBuffer("@%s", name), value);
-    }
-
-public:
-    virtual IPTree* queryTree() const
-    {
-        return tree.get();
-    }
-
-protected:
-    Owned<IPTree> tree;
-    IPTree* active = nullptr;
-};
-
-IEventPTreeCreator* createEventPTreeCreator()
-{
-    class Creator : public CInterfaceOf<IEventPTreeCreator>
-    {
-    public:
-        virtual IEventVisitor& queryVisitor() override
-        {
-            return *visitor;
-        }
-
-        virtual IPTree* queryTree() const override
-        {
-            return visitor->queryTree();
-        }
-
-        Creator()
-        {
-            visitor.setown(new CPTreeEventVisitor);
-        }
-
-    private:
-        Owned<CPTreeEventVisitor> visitor;
-    };
-    return new Creator;
-}
-
-// Abstract extension of CDumpEventVisitor that writes visited data to an output stream. Subclasses
-// are responsible for formatting the data in the desired format.
-class CDumpStreamEventVisitor : public CDumpEventVisitor
-{
-protected: // new abstract method(s)
-    virtual void recordAttribute(EventAttr id, const char* name, const char* value, bool quoted) = 0;
-
-public:
-    CDumpStreamEventVisitor(IBufferedSerialOutputStream& _out)
-        : out(&_out)
-    {
-    }
-
-protected:
-    void dump(bool eoln)
-    {
-        out->put(markup.length(), markup.str());
-        markup.clear();
-        if (eoln)
-            out->put(1, "\n");
-    }
-
-    bool inHeader() const { return State::Header == state; }
-    bool inEvents() const { return State::Events == state; }
-
-protected:
-    enum class State : byte
-    {
-        Idle,
-        Header,
-        Events,
-        Footer,
-    };
-    State state{State::Idle};
-    Linked<IBufferedSerialOutputStream> out;
-    StringBuffer markup;
-};
-
-// Concrete extension of CDumpStreamEventVisitor that writes visited data as unstructured text.
-// One line of text is produced for each event and attribute visited, plus additional lines for
-// the file name, version, and number of bytes read from the file.
-class CDumpTextEventVisitor : public CDumpStreamEventVisitor
-{
-public:
-    virtual bool visitFile(const char* filename, uint32_t version) override
-    {
-        state = State::Header;
-        doVisitHeader(filename, version);
-        return true;
-    }
-
-    virtual Continuation visitEvent(EventType id) override
-    {
-        if (inHeader())
-        {
-            closeElement();
-            state = State::Events;
-        }
-        openElement(queryEventName(id));
-        doVisitEvent(id);
-        return visitContinue;
-    }
-
-    virtual bool departEvent() override
-    {
-        closeElement();
-        return true;
-    }
-
-    virtual void departFile(uint32_t bytesRead) override
-    {
-        if (inHeader())
-            closeElement();
-        state = State::Footer;
-        doVisitFooter(bytesRead);
-        closeElement();
-    }
-
-protected:
-    virtual void recordAttribute(EventAttr id, const char* name, const char* value, bool quoted) override
-    {
-        markup.append("attribute: ").append(name).append(" = ");
-        if (quoted)
-            markup.append('\'');
-        markup.append(value);
-        if (quoted)
-            markup.append('\'');
-        markup.append('\n');
-    }
-
-public:
-    using CDumpStreamEventVisitor::CDumpStreamEventVisitor;
-
-protected:
-    void openElement(const char* name)
-    {
-        markup.append("event: ").append(name).append('\n');
-    }
-
-    void closeElement()
-    {
-        dump(false);
-    }
-};
-
-IEventVisitor* createDumpTextEventVisitor(IBufferedSerialOutputStream& out)
-{
-    return new CDumpTextEventVisitor(out);
-}
-
-// Extension of CDumpStreamEventVisitor that outputs XML-formatted text. Header and Footer elements
-// synthesized to hold the file name, version, and number of bytes read from the file.
-class CDumpXMLEventVisitor : public CDumpStreamEventVisitor
-{
-public:
-    virtual bool visitFile(const char* filename, uint32_t version) override
-    {
-        state = State::Header;
-        openElement(DUMP_STRUCTURE_ROOT, true);
-        openElement(DUMP_STRUCTURE_HEADER);
-        doVisitHeader(filename, version);
-        return true;
-    }
-
-    virtual Continuation visitEvent(EventType id) override
-    {
-        if (inHeader())
-        {
-            closeElement();
-            state = State::Events;
-        }
-        openElement(DUMP_STRUCTURE_EVENT);
-        doVisitEvent(id);
-        return visitContinue;
-    }
-
-    virtual bool departEvent() override
-    {
-        closeElement();
-        return true;
-    }
-
-    virtual void departFile(uint32_t bytesRead) override
-    {
-        if (inHeader())
-            closeElement();
-        state = State::Footer;
-        openElement(DUMP_STRUCTURE_FOOTER);
-        doVisitFooter(bytesRead);
-        closeElement();
-        closeElement(DUMP_STRUCTURE_ROOT);
-    }
-
-protected:
-    virtual void recordAttribute(EventAttr id, const char* name, const char* value, bool quoted) override
-    {
-        appendXMLAttr(markup, name, value);
-    }
-
-public:
-    using CDumpStreamEventVisitor::CDumpStreamEventVisitor;
-
-    void openElement(const char* name, bool complete = false)
-    {
-        markup.pad(2 * indentLevel);
-        indentLevel++;
-        appendXMLOpenTag(markup, name, nullptr, complete);
-        if (complete)
-            markup.append('\n');
-    }
-
-    void closeElement()
-    {
-        markup.append("/>");
-        indentLevel--;
-        dump(true);
-    }
-
-    void closeElement(const char* name)
-    {
-        appendXMLCloseTag(markup, name);
-        indentLevel--;
-        dump(true);
-    }
-
-private:
-    unsigned indentLevel{0};
-};
-
-IEventVisitor* createDumpXMLEventVisitor(IBufferedSerialOutputStream& out)
-{
-    return new CDumpXMLEventVisitor(out);
-}
-
-// Extension of CDumpStreamEventVisitor that outputs JSON-formatted text. Header and Footer objects
-// are synthesized to hold the file name, version, and number of bytes read from the file.
-class CDumpJSONEventVisitor : public CDumpStreamEventVisitor
-{
-public:
-    virtual bool visitFile(const char* filename, uint32_t version) override
-    {
-        state = State::Header;
-        openElement();
-        openElement(DUMP_STRUCTURE_HEADER);
-        doVisitHeader(filename, version);
-        return true;
-    }
-
-    virtual Continuation visitEvent(EventType id) override
-    {
-        if (inHeader())
-        {
-            closeElement();
-            state = State::Events;
-        }
-        if (firstEvent)
-        {
-            openArray(DUMP_STRUCTURE_EVENT);
-            openElement();
-            firstEvent = false;
-        }
-        else
-            openElement();
-        doVisitEvent(id);
-        return visitContinue;
-    }
-
-    virtual bool departEvent() override
-    {
-        closeElement();
-        return true;
-    }
-
-    virtual void departFile(uint32_t bytesRead) override
-    {
-        if (inHeader())
-            closeElement();
-        else if (inEvents())
-            closeArray();
-        openElement(DUMP_STRUCTURE_FOOTER);
-        doVisitFooter(bytesRead);
-        closeElement();
-        closeElement(true);
-    }
-
-protected:
-    virtual void recordAttribute(EventAttr id, const char* name, const char* value, bool quoted) override
-    {
-        conditionalDelimiter();
-        indent();
-        appendJSONStringValue(markup, name, value, true, quoted);
-    }
-
-public:
-    using CDumpStreamEventVisitor::CDumpStreamEventVisitor;
-
-protected:
-    inline void openElement()
-    {
-        openElement(nullptr);
-    }
-
-    void openElement(const char* name)
-    {
-        openContainer(name, "{ ");
-    }
-
-    void openArray(const char* name)
-    {
-        openContainer(name, "[ ");
-    }
-
-    void closeArray()
-    {
-        closeContainer("]");
-    }
-
-    void closeElement(bool done = false)
-    {
-        closeContainer("}");
-        dump(done);
-    }
-
-    void openContainer(const char* name, const char* token)
-    {
-        conditionalDelimiter();
-        indent();
-        if (!isEmptyString(name))
-            appendJSONName(markup, name);
-        markup.append(token);
-        firstFlags.push_back(true);
-    }
-
-    void closeContainer(const char* token)
-    {
-        firstFlags.pop_back();
-        indent();
-        markup.append(token);
-    }
-
-    void conditionalDelimiter()
-    {
-        if (!firstFlags.empty())
-        {
-            if (!firstFlags.back())
-                markup.append(",");
-            else
-                firstFlags.back() = false;
-        }
-    }
-
-    void indent()
-    {
-        markup.append('\n');
-        if (!firstFlags.empty())
-            markup.pad(2 * (firstFlags.size()));
-    }
-
-private:
-    std::vector<bool> firstFlags;
-    bool firstEvent{true};
-};
-
-IEventVisitor* createDumpJSONEventVisitor(IBufferedSerialOutputStream& out)
-{
-    return new CDumpJSONEventVisitor(out);
-}
-
-// Extension of CDumpStreamEventVisitor that outputs YAML-formatted text. Header and Footer objects
-// are synthesized to hold the file name, version, and number of bytes read from the file.
-class CDumpYAMLEventVisitor : public CDumpStreamEventVisitor
-{
-public:
-    virtual bool visitFile(const char* filename, uint32_t version) override
-    {
-        state = State::Header;
-        openElement(DUMP_STRUCTURE_HEADER);
-        doVisitHeader(filename, version);
-        return true;
-    }
-
-    virtual Continuation visitEvent(EventType id) override
-    {
-        if (inHeader())
-        {
-            closeElement();
-            openElement(DUMP_STRUCTURE_EVENT);
-            state = State::Events;
-        }
-        else
-            openElement(nullptr);
-        doVisitEvent(id);
-        return visitContinue;
-    }
-
-    virtual bool departEvent() override
-    {
-        closeElement();
-        return true;
-    }
-
-    virtual void departFile(uint32_t bytesRead) override
-    {
-        if (inHeader())
-            closeElement();
-        state = State::Footer;
-        openElement(DUMP_STRUCTURE_FOOTER);
-        doVisitFooter(bytesRead);
-        closeElement();
-    }
-
-protected:
-    virtual void recordAttribute(EventAttr id, const char* name, const char* value, bool quoted) override
-    {
-        indent();
-        markup.append(name).append(": ").append(value).append('\n');
-    }
-
-public:
-    using CDumpStreamEventVisitor::CDumpStreamEventVisitor;
-
-    void openElement(const char* name)
-    {
-        if (!isEmptyString(name))
-        {
-            indent();
-            markup.append(name).append(":\n");
-        }
-        indentLevel++;
-        firstProp = true;
-    }
-
-    void closeElement()
-    {
-        indentLevel--;
-        dump(false);
-    }
-
-    void indent()
-    {
-        if (indentLevel)
-        {
-            uint8_t limit = indentLevel;
-            if (inEvents() && firstProp)
-                limit -= 1;
-            markup.pad(2 * limit);
-            if (inEvents() && firstProp)
-            {
-                markup.append("- ");
-                firstProp = false;
-            }
-        }
-    }
-
-protected:
-    bool firstProp{true};
-    uint8_t indentLevel{0};
-};
-
-IEventVisitor* createDumpYAMLEventVisitor(IBufferedSerialOutputStream& out)
-{
-    return new CDumpYAMLEventVisitor(out);
-}
-
-class CDumpCSVEventVisitor : public CDumpStreamEventVisitor
-{
-public:
-    virtual bool visitFile(const char* filename, uint32_t version) override
-    {
-
-        state = State::Header;
-        encodeCSVColumn(markup, "EventName");
-        for (unsigned a = EvAttrNone + 1; a < EvAttrMax; a++)
-        {
-            markup.append(",");
-            encodeCSVColumn(markup, queryEventAttributeName(EventAttr(a)));
-        }
-        dump(true);
-        return true;
-    }
-
-    virtual Continuation visitEvent(EventType id) override
-    {
-        if (inHeader())
-            state = State::Events;
-        encodeCSVColumn(markup, queryEventName(id));
-        return visitContinue;
-    }
-
-    virtual bool departEvent() override
-    {
-        for (unsigned a = EvAttrNone + 1; a < EvAttrMax; a++)
-        {
-            markup.append(",");
-            if (row[a].get())
-                encodeCSVColumn(markup, row[a].get());
-        }
-        dump(true);
-        // Prepare for the next event by clearing only those row values set by the departed event.
-        // Header attributes are not cleared so that they can be output for each event.
-        for (unsigned a = EvAttrNone + 1; a < EvAttrMax; a++)
-        {
-            if (row[a].get() && !headerAttrs.count(EventAttr(a)))
-                row[a].clear();
-        }
-        return true;
-    }
-
-    virtual void departFile(uint32_t bytesRead) override
-    {
-    }
-
-protected:
-    virtual void recordAttribute(EventAttr id, const char* name, const char* value, bool quoted) override
-    {
-        if (id != EvAttrNone)
-        {
-            row[id].set(value);
-            if (inHeader())
-                headerAttrs.insert(id);
-        }
-    }
-
-public:
-    using CDumpStreamEventVisitor::CDumpStreamEventVisitor;
-
-private:
-    StringAttr row[EvAttrMax];
-    std::set<EventAttr> headerAttrs;
-};
-
-IEventVisitor* createDumpCSVEventVisitor(IBufferedSerialOutputStream& out)
-{
-    return new CDumpCSVEventVisitor(out);
-}
-
-class CEventFilter : public CInterfaceOf<IEventFilter>
-{
-protected:
-    class StringMatchHelper
-    {
-    private:
-        using Patterns = std::set<std::string>;
-        Patterns patterns;
-
-    public:
-        bool matches(const char* value) const
-        {
-            for (const std::string& pattern : patterns)
-            {
-                if (WildMatch(value, pattern.c_str(), false))
-                    return true;
-            }
-            return false;
-        }
-
-        bool acceptToken(const char* token)
-        {
-            patterns.insert(token);
-            return true;
-        }
-    };
-
-    struct FilterTerm : public CInterface
-    {
-        virtual bool matches(EventType evt, const char* value) const { return true; }
-        virtual bool matches(EventType evt, bool value) const { return true; }
-        virtual bool matches(EventType evt, __uint64 value) const { return true; }
-        virtual bool accept(const char* values)
-        {
-            StringArray tokens;
-            tokens.appendList(values, ",", true);
-            ForEachItemIn(i, tokens)
-            {
-                if (!acceptToken(tokens.item(i)))
-                    return false;
-            }
-            return true;
-        }
-        virtual bool acceptToken(const char* token) = 0;
-    };
-
-    struct StringFilterTerm : public FilterTerm
-    {
-        StringMatchHelper helper;
-        bool matches(EventType evt, const char* value) const override
-        {
-            return helper.matches(value);
-        }
-
-        bool acceptToken(const char* token) override
-        {
-            return helper.acceptToken(token);
-        }
-    };
-
-    struct BoolFilterTerm : public FilterTerm
-    {
-        bool accepted[2]{false, false};
-
-        bool matches(EventType evt, bool value) const override
-        {
-            return accepted[value];
-        }
-
-        bool acceptToken(const char* token) override
-        {
-            bool value = strToBool(token);
-            accepted[value] = true;
-            return true;
-        }
-    };
-
-    struct UnsignedFilterTerm : public FilterTerm
-    {
-        using Accepted = std::set<std::pair<__uint64, __uint64>>;
-        Accepted accepted;
-
-        bool matches(EventType evt, __uint64 value) const override
-        {
-            for (const std::pair<__uint64, __uint64>& range : accepted)
-            {
-                if (range.first <= value && value <= range.second)
-                    return true;
-            }
-            return false;
-        }
-
-        bool acceptToken(const char* token) override
-        {
-            char* next = nullptr;
-            const char* delim = strchr(token, '-');
-            if (!delim)
-            {
-                __uint64 first = strtoull(token, &next, 0);
-                if (*next == '\0')
-                    return acceptRange(first, first);
-            }
-            else if (delim == token)
-            {
-                __uint64 last = strtoull(delim + 1, &next, 0);
-                if (*next == '\0')
-                    return acceptRange(0, last);
-            }
-            else
-            {
-                __uint64 first = strtoull(token, &next, 0);
-                while (isspace(*next))
-                    next++;
-                if (*next == '-')
-                {
-                    if (!*(next + 1))
-                        return acceptRange(first, UINT64_MAX);
-                    __uint64 last = strtoull(next + 1, &next, 0);
-                    if (*next == '\0')
-                        return acceptRange(first, last);
-                }
-            }
-            return false;
-        }
-
-        bool acceptRange(__uint64 first, __uint64 last)
-        {
-            if (first <= last)
-                return accepted.insert(std::make_pair(first, last)).second;
-            return false;
-        }
-    };
-
-    struct FileIdFilterTerm : public UnsignedFilterTerm
-    {
-        bool matches(EventType evt, const char* value) const override
-        {
-            if (MetaFileInformation == evt)
-            {
-                // Test the presumed EvAttrPath value for a pattern match. On a match, the current
-                // meta file id is added to the set of matched paths.
-                //
-                // Assumes that EvAttrFileId is recorded before EvAttrPath.
-                if (curMetaId && helper.matches(value))
-                    matchedPaths.insert(curMetaId);
-            }
-            return true;
-        }
-        bool matches(EventType evt, __uint64 value) const override
-        {
-            if (MetaFileInformation == evt)
-            {
-                // Cache the presumed EvAttrFileId value for use when matching EvAttrPath.
-                curMetaId = value;
-                return true;
-            }
-            return matchedPaths.count(value) ||  UnsignedFilterTerm::matches(evt, value);
-        }
-        bool acceptToken(const char* token) override
-        {
-            if (UnsignedFilterTerm::acceptToken(token))
-                return true;
-            // Token not a number or range; assume it is a path pattern
-            return helper.acceptToken(token);
-        }
-
-        StringMatchHelper helper;
-        mutable std::set<__uint64> matchedPaths;
-        mutable __uint64 curMetaId{0};
-    };
-
-    struct TimestampFilterTerm : public UnsignedFilterTerm
-    {
-        bool acceptToken(const char* token) override
-        {
-            // Digits and hyphen only is presumed to be a nanosecond timestamp range.
-            std::string tmp(token);
-            if (tmp.find_first_not_of("0123456789-") == std::string::npos)
-                return UnsignedFilterTerm::acceptToken(token);
-
-            // All else is presumed to be a date/time string or a date/time range.
-            __uint64 first = 0, last = UINT64_MAX;
-            CDateTime dt;
-            const char* delim = nullptr;
-            if (*token == '-') // range starting at 0
-            {
-                dt.setString(token + 1, &delim);
-                if (delim && *delim) // bad content
-                    return false;
-                last = dt.getTimeStampNs();
-            }
-            else // single value or range not starting at 0
-            {
-                dt.setString(token, &delim);
-                first = dt.getTimeStampNs();
-                if (!delim) // single value
-                    last = first;
-                else
-                {
-                    skipSpaces(delim);
-                    if ('-' == *delim) // a range...
-                    {
-                        token = delim + 1;
-                        if (*token) // with bounded end
-                        {
-                            skipSpaces(token);
-                            delim = nullptr;
-                            dt.setString(token, &delim);
-                            if (delim && *delim) // bad content
-                                return false;
-                            last = dt.getTimeStampNs();
-                        }
-                        else // with unbounded end
-                            last = first;
-                    }
-                    else // bad content
-                        return false;
-                }
-            }
-            return acceptRange(first, last);
-        }
-
-        void skipSpaces(const char*& token)
-        {
-            while (isspace(*token))
-                token++;
-        }
-    };
-
-    struct CachedAttrValue
-    {
-        StringBuffer s;
-        __uint64 u{0};
-        bool b{false};
-    };
-
-public: // IEventVisitor
-    virtual bool visitFile(const char* filename, uint32_t version) override
-    {
-        return target->visitFile(filename, version);
-    }
-
-    virtual Continuation visitEvent(EventType type) override
-    {
-        if (!acceptedEvents.empty() && !acceptedEvents.count(type))
-        {
-            // MetaFileInformation cannot be skipped yet if there is a file id filter term. It
-            // must be skipped from departEvent after the any path evaluations are completed.
-            if (type != MetaFileInformation || !terms[EvAttrFileId])
-                return visitSkipEvent;
-        }
-        cacheSequence.clear();
-        currentEvent = type;
-        return visitContinue;
-    }
-
-    virtual Continuation visitAttribute(EventAttr id, const char* value) override
-    {
-        if (EventNone == currentEvent) // file header/footer attributes are never filtered
-            return visitContinue;
-        if (terms[id] && !terms[id]->matches(currentEvent, value))
-            return visitSkipEvent;
-        cachedAttrValues[id].s.set(value);
-        cacheSequence.push_back(id);
-        return visitContinue;
-    }
-
-    virtual Continuation visitAttribute(EventAttr id, bool value) override
-    {
-        if (EventNone == currentEvent) // file header/footer attributes are never filtered
-            return visitContinue;
-        if (terms[id] && !terms[id]->matches(currentEvent, value))
-            return visitSkipEvent;
-        cachedAttrValues[id].b = value;
-        cacheSequence.push_back(id);
-        return visitContinue;
-    }
-
-    virtual Continuation visitAttribute(EventAttr id, __uint64 value) override
-    {
-        if (EventNone == currentEvent) // file header/footer attributes are never filtered
-            return visitContinue;
-        if (terms[id] && !terms[id]->matches(currentEvent, value))
-            return visitSkipEvent;
-        cachedAttrValues[id].u = value;
-        cacheSequence.push_back(id);
-        return visitContinue;
-    }
-
-    virtual bool departEvent() override
-    {
-        // Skip any event that could not be skipped during visitEvent
-        if (!acceptedEvents.empty() && !acceptedEvents.count(currentEvent))
-        {
-            currentEvent = EventNone;
-            cacheSequence.clear();
-            return true;
-        }
-
-        // Tell the target about the current event, obeying its continuation directive.
-        switch (target->visitEvent(currentEvent))
-        {
-        case visitSkipEvent:
-            currentEvent = EventNone;
-            cacheSequence.clear();
-            return true;
-        case visitSkipFile:
-            return false;
-        case visitContinue:
-            break;
-        }
-
-        // Tell the target about the event's attributes, obeying its continuation directives.
-        Continuation continuation = visitContinue;
-        for (EventAttr id : cacheSequence)
-        {
-            switch (attrInformation[id].type)
-            {
-            case EATbool:
-                continuation = target->visitAttribute(id, cachedAttrValues[id].b);
-                break;
-            case EATtraceid:
-            case EATstring:
-                continuation = target->visitAttribute(id, cachedAttrValues[id].s);
-                break;
-            case EATnone:
-            case EATmax:
-                throw makeStringExceptionV(-1, "event attribute id %d has invalid type %d", int(id), int(attrInformation[id].type));
-            default:
-                continuation = target->visitAttribute(id, cachedAttrValues[id].u);
-                break;
-            }
-            switch (continuation)
-            {
-            case visitSkipEvent:
-                currentEvent = EventNone;
-                cacheSequence.clear();
-                return true;
-            case visitSkipFile:
-                return false;
-            case visitContinue:
-                break;
-            }
-        }
-
-        // Tell the target that the event is complete, obeying its continuation directive.
-        return target->departEvent();
-    }
-
-    virtual void departFile(uint32_t bytesRead) override
-    {
-        target->departFile(bytesRead);
-    }
-
-public:
-    virtual bool acceptEvent(EventType type) override
-    {
-        return (acceptedEvents.insert(type).second);
-    }
-
-    virtual bool acceptEvents(EventContext context) override
-    {
-        bool changed = false;
-        for (byte type = EventNone + 1; type < EventMax; type++)
-        {
-            if (queryEventContext(EventType(type)) == context)
-            {
-                if (acceptedEvents.insert(EventType(type)).second && !changed)
-                    changed = true;
-            }
-        }
-        return changed;
-    }
-
-    virtual bool acceptEvents(const char* events) override
-    {
-        StringArray tokens;
-        tokens.appendList(events, ",", true);
-        bool changed = false;
-        ForEachItemIn(i, tokens)
-        {
-            EventContext context = queryEventContext(tokens.item(i));
-            if (context != EventCtxMax)
-            {
-                if (acceptEvents(context) && !changed)
-                    changed = true;
-                continue;
-            }
-            EventType type = queryEventType(tokens.item(i));
-            if (type != EventNone)
-            {
-                if (acceptEvent(type) && !changed)
-                    changed = true;
-                continue;
-            }
-        }
-        return changed;
-    }
-
-    virtual bool acceptAttribute(EventAttr id, const char* values) override
-    {
-        if (EvAttrFileId == id)
-        {
-            // Special case to join EventIndexLookup, EventIndexLoad, and EventIndexEvtion with
-            // MetaFileInformation by the common EvAttrFileId attribute. The result allows the
-            // index events to be filtered by path using the file ID.
-            //
-            // MetaFileInformation events cannot be filtered by file ID.
-            FilterTerm* term = ensureTerm<FileIdFilterTerm>(id);
-            if (!terms[EvAttrPath])
-                terms[EvAttrPath].set(term);
-            else if (terms[EvAttrPath] != term)
-                throw makeStringException(-1, "event attribute EvAttrPath has a conflicting filter term for EvAttrFileId");
-            return term->accept(values);
-        }
-        else
-        {
-            switch (attrInformation[id].type)
-            {
-            case EATtraceid:
-            case EATstring:
-                return ensureTerm<StringFilterTerm>(id)->accept(values);
-            case EATbool:
-                return ensureTerm<BoolFilterTerm>(id)->accept(values);
-            case EATu1:
-            case EATu2:
-            case EATu4:
-            case EATu8:
-                return ensureTerm<UnsignedFilterTerm>(id)->accept(values);
-            case EATtimestamp:
-                return ensureTerm<TimestampFilterTerm>(id)->accept(values);
-            case EATnone:
-            case EATmax:
-                throw makeStringExceptionV(-1, "event attribute id %d has an invalid type %d", int(id), int(attrInformation[id].type));
-            default:
-                throw makeStringExceptionV(-1, "event attribute id %d has an unknown type %d", int(id), int(attrInformation[id].type));
-            }
-        }
-    }
-
-    virtual void setTarget(IEventVisitor& visitor) override
-    {
-        target.set(&visitor);
-    }
-
-protected:
-    template <typename term_type_t>
-    FilterTerm* ensureTerm(EventAttr id)
-    {
-        if (id <= EvAttrNone || id >= EvAttrMax)
-            throw makeStringExceptionV(-1, "event attribute id %d out of range", int(id));
-        if (!terms[id])
-            terms[id].setown(new term_type_t);
-#if defined(_DEBUG)
-        else if (dynamic_cast<term_type_t*>(terms[id].get()) == nullptr)
-            throw makeStringExceptionV(-1, "event attribute id %d has a different type of filter term", int(id));
-#endif
-        return terms[id].get();
-    }
-
-    bool isEventAccepted(EventType type) const
-    {
-        return (acceptedEvents.empty() || acceptedEvents.count(type));
-    }
-
-protected:
-    using AcceptedEvents = std::unordered_set<EventType>;
-    Linked<IEventVisitor> target;
-    Owned<FilterTerm> terms[EvAttrMax];
-    CachedAttrValue cachedAttrValues[EvAttrMax];
-    using CacheSequence = std::vector<EventAttr>;
-    CacheSequence cacheSequence;
-    AcceptedEvents acceptedEvents;
-    EventType currentEvent{EventNone};
-};
-
-IEventFilter* createEventFilter()
-{
-    return new CEventFilter;
 }
 
 // GH->TK
