+++ system/jlib/jevent.hpp
@@ -58,6 +58,7 @@ enum EventContext : byte
 };
 
 extern jlib_decl EventContext queryEventContext(EventType event);
+extern jlib_decl EventContext queryEventContext(const char* name);
 
 // The attributes that can be associated with each event
 // The order should not be changed, or items removed.  New values should always be appended before EvAttrMax
@@ -85,10 +86,26 @@ enum EventAttr : byte
     EvAttrMax
 };
 
+// The different data types of attributes that can be associated with an event.
+enum EventAttrType
+{
+    EATnone,
+    EATbool,
+    EATu1,
+    EATu2,
+    EATu4,
+    EATu8,
+    EATtimestamp,
+    EATstring,
+    EATtraceid,     // should probably delete...
+    EATmax
+};
+
 extern jlib_decl EventType queryEventType(const char* name);
 extern jlib_decl const char * queryEventName(EventType event);
 extern jlib_decl EventAttr queryEventAttribute(const char* name);
 extern jlib_decl const char * queryEventAttributeName(EventAttr attr);
+extern jlib_decl EventAttrType queryEventAttributeType(EventAttr attr);
 
 struct jlib_decl EventRecordingSummary
 {
@@ -271,82 +288,11 @@ interface IEventVisitor : extends IInterface
     virtual void departFile(uint32_t bytesRead) = 0;
 };
 
-// Get a visitor that streams visited event data in JSON format.
-extern jlib_decl IEventVisitor* createDumpJSONEventVisitor(IBufferedSerialOutputStream& out);
-
-// Get a visitor that streams visited event data in a flat text format.
-extern jlib_decl IEventVisitor* createDumpTextEventVisitor(IBufferedSerialOutputStream& out);
-
-// Get a visitor that streams visited event data in XML format.
-extern jlib_decl IEventVisitor* createDumpXMLEventVisitor(IBufferedSerialOutputStream& out);
-
-// Get a visitor that streams visited event data in YAML format.
-extern jlib_decl IEventVisitor* createDumpYAMLEventVisitor(IBufferedSerialOutputStream& out);
-
-// Get a visitor that streams visited event data in CSV format.
-extern jlib_decl IEventVisitor* createDumpCSVEventVisitor(IBufferedSerialOutputStream& out);
-
-// Encapsulation of a visitor that stores visited event data in a property tree and
-// access to the tree.
-interface IEventPTreeCreator : extends IInterface
-{
-    virtual IEventVisitor& queryVisitor() = 0;
-    virtual IPTree* queryTree() const = 0;
-};
-
-// Get an event property tree creator.
-extern jlib_decl IEventPTreeCreator* createEventPTreeCreator();
-
-// Extension of IEventVisitor that supports filtering of visited files and events. Implementations
-// will decorate another visitor, forwarding all visits not blocked by the specified constraints to
-// the decorated visitor.
-//
-// A file blocked by a version constraint must visit and immediately depart the file.
-//
-// An event blocked by a type or context constraint must not be forwarded to the decorated visitor.
-//
-// An event blocked by an attribute constraint must not be forwarded to the decorated visitor.
-// Events that do not include a constrained attribute are not blocked.
-interface IEventFilter : extends IEventVisitor
-{
-    // Filter on a single event type. All events are accepted by default.
-    virtual bool acceptEvent(EventType type) = 0;
-    // Filter on all events of a given context. All events are accepted by default.
-    virtual bool acceptEvents(EventContext context) = 0;
-    // Filter on a comma-delimited list of event type names and/or event context names. All events
-    // are accepted by default.
-    virtual bool acceptEvents(const char* types) = 0;
-    // Filter on a comma-delimited list of attribute value tokens. The internal type of the given
-    // attribute is used to determine token processing requirements.
-    // - Integral token options are:
-    //   - #: a single numeric value
-    //   - #-#: a range of numeric values, bounded on both ends
-    //   - #-: a range of numeric values, bounded only on the lower end
-    //   - -#: a range of numeric values, bounded only on the upper end
-    // - String token options include exact matches and regular expressions.
-    // - Boolean token options are text representations of true and false recognized by strToBool.
-    //   Specification of multiple Boolean values is unnecessary - either they will be repetitive
-    //   or will cancel each other out.
-    //
-    // Supported special cases include:
-    // - Timestamps may be given using a standard date/time format, such as "2025-01-01T00:00:00".
-    // - A filter for the integral EvAttrFileId may include string tokens that will be applied to
-    //   a corresponding EvAttrPath attribute previously observed in a MetaFileInformation event.
-    virtual bool acceptAttribute(EventAttr attr, const char* values) = 0;
-
-    // Install the recipient of unfiltered visits. `readEvents` will call this method before
-    // beginning visitation.
-    virtual void setTarget(IEventVisitor& visitor) = 0;
-};
-
-// Obtain a new instance of a standard event filter.
-extern jlib_decl IEventFilter* createEventFilter();
-
 // Opens and parses a single binary event data file. Parsed data is passed to the given visitor
 // until parsing completes or the visitor requests it to stop.
 //
 // Exceptions are thrown on error. False is returned if parsing was stopped prematurely. True is
 // returned if all data was parsed successfully.
-extern jlib_decl bool readEvents(const char* filename, IEventVisitor & visitor, IEventFilter* filter);
+extern jlib_decl bool readEvents(const char* filename, IEventVisitor & visitor);
 
 #endif
