+++ testing/unittests/jlibtests2.cpp
@@ -30,7 +30,7 @@
 #include "jsem.hpp"
 #include "jfile.hpp"
 #include "jevent.hpp"
-
+#include "eventdump.h"
 #include "unittests.hpp"
 
 
@@ -342,7 +342,7 @@ attribute: bytesRead = 21
             StringBuffer out;
             Owned<IEventVisitor> visitor = createVisitor(out);
             CPPUNIT_ASSERT(visitor.get());
-            CPPUNIT_ASSERT(readEvents("eventtrace.evt", *visitor, nullptr));
+            CPPUNIT_ASSERT(readEvents("eventtrace.evt", *visitor));
             CPPUNIT_ASSERT_EQUAL_STR(expect, out.str());
             DBGLOG("Raw size = %llu, File size = %llu", summary.rawSize, summary.totalSize);
         }
@@ -380,7 +380,7 @@ attribute: bytesRead = 76
             StringBuffer out;
             Owned<IEventVisitor> visitor = createVisitor(out);
             CPPUNIT_ASSERT(visitor.get());
-            CPPUNIT_ASSERT(readEvents("eventtrace.evt", *visitor, nullptr));
+            CPPUNIT_ASSERT(readEvents("eventtrace.evt", *visitor));
             CPPUNIT_ASSERT_EQUAL_STR(expect, out.str());
             DBGLOG("Raw size = %llu, File size = %llu", summary.rawSize, summary.totalSize);
         }
@@ -428,7 +428,7 @@ attribute: bytesRead = 161
             StringBuffer out;
             Owned<IEventVisitor> visitor = createVisitor(out);
             CPPUNIT_ASSERT(visitor.get());
-            CPPUNIT_ASSERT(readEvents("eventtrace.evt", *visitor, nullptr));
+            CPPUNIT_ASSERT(readEvents("eventtrace.evt", *visitor));
             CPPUNIT_ASSERT_EQUAL_STR(expect, out.str());
             DBGLOG("Raw size = %llu, File size = %llu", summary.rawSize, summary.totalSize);
         }
